Moved spatial types into EsqlDataTypes
